master-degree/laboratorio-progettazione/backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==3.3.9
autopep8==2.3.2
bandit==1.8.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==41.0.0
dill==0.3.9
dnspython==2.7.0
dodgy==0.2.1
email_validator==2.2.0
Faker==37.0.1
fastapi==0.115.11
fastapi-cli==0.0.7
flake8==7.1.2
flake8-polyfill==1.0.2
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
orjson==3.10.16
packaging==24.2
passlib==1.7.4
pbr==6.1.1
pdoc==15.0.1
pep8-naming==0.10.0
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
prospector==1.15.3
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydocstyle==6.3.0
pyflakes==3.2.0
Pygments==2.19.1
pylint==3.3.6
pylint-celery==0.3
pylint-django==2.6.1
pylint-flask==0.6
pylint-plugin-utils==0.8.2
PyNaCl==1.5.0
pyotp==2.9.0
pytest==8.3.5
pytest-mock==3.14.0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
qrcode==8.2.0
requests==2.32.3
requirements-detector==1.3.2
rich==13.9.4
rich-toolkit==0.14.1
semver==3.0.4
setoptconf-tmp==0.3.1
setuptools==78.0.2
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
SQLAlchemy==2.0.39
starlette==0.46.1
stevedore==5.4.1
toml==0.10.2
tomlkit==0.13.2
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1